+++ src/client/debugger/resources.tsx
@@ -421,7 +421,7 @@ export const GfxResourcesView: FunctionComponent<{
 		const copperScreens: { screen: IScreen; frames: number[] }[] = [];
 		// dupecheck copper screens from all frames
 		for(let i = 0; i < MODELS.length; i++) {
-			const copperScreen = GetScreenFromCopper(MODELS[i].copper);
+			const copperScreen = GetScreenFromCopper(MODELS[i].copper, MODELS[0].amiga.chipsetFlags);
 			const dupe = copperScreens.findIndex((cs) => JSON.stringify(cs.screen) === JSON.stringify(copperScreen));
 			if(dupe === -1)
 				copperScreens.push({ screen: copperScreen, frames: [i + 1] });
@@ -488,7 +488,7 @@ export const GfxResourcesView: FunctionComponent<{
 		};
 		palettes.push({ resource: copperResource, frame, palette: copperPalette });
 
-		const customRegs = GetCustomRegsAfterDma(MODELS[frame].amiga.customRegs, MODELS[frame].amiga.dmacon, MODELS[frame].amiga.dmaRecords, CpuCyclesToDmaCycles(time));
+		const customRegs = GetCustomRegsAfterDma(MODELS[frame].amiga.customRegs, MODELS[frame].amiga.dmaRecords, CpuCyclesToDmaCycles(time));
 		const customRegsPalette = GetPaletteFromCustomRegs(new Uint16Array(customRegs));
 		const customRegsResource: GfxResource = {
 			address: CustomRegisters.getCustomAddress("COLOR00"),